Support for building LLVM 3.7 and 3.8 on CentOS 5
tar on CentOS 5.8 does not support the J option. As a workaround, use
the standalone xzcat utility.

CentOS 5.8 does not have the perf_event.h header, needed by compiler-rt.
We do not need compiler-rt for cross-compiled IR so it is still useful
to enable LLVM without compiler-rt on CentOS5.8. The various LLVM
sanitizers, including ASAN require compiler-rt so these will not work on
CentOS 5.8.

Change-Id: Iad649363fc64f30e0f17a9d51d7694a6fc00bc12