Support for building LLVM 3.7 and 3.8 on CentOS 5
tar on CentOS 5.8 does not support the J option. As a workaround, use
the standalone xzcat utility.

CentOS 5.8 does not have the perf_event.h header, needed by compiler-rt.
We can do without compiler-rt. It is needed for the various sanitisers,
and also provides an alternative to the gcc runtime C/C++ runtime libraries.

Change-Id: Iad649363fc64f30e0f17a9d51d7694a6fc00bc12