IMPALA-3191: toolchain components should use toolchain cmake
The toolchain doesn't build on sles12. It appears to be a
cmake bug [1] though it hasn't been fixed yet upstream.

The cmake bug description mentions this workaround patching
cmake to avoid calling POSIX select(), and we should be
using the toolchain cmake to build toolchain components.

This also reverts the previous attempt to mitigate the issue
by avoiding bootstrapping cmake in parallel. That seemed to
help avoid the issue during bootstrapping, but the hang
could still occur later as well.

1: https://cmake.org/Bug/view.php?id=15873

Change-Id: I8a31098de42b4d5cb93c12bb2e054eadb8b8c72a