Add Kudu using parcels
As a quick way to get started, Kudu will be obtained from the Cloudera
parcel repository. Kudu parcels are currently only available for a few
operating systems. For operating systems that don't have a parcel, a
stubbed client is generated. The stub client only provides the same
symbols that the real client would provide. The stub is useful for
building but it is non-functional and its methods should not be called.

In a later version, Kudu will be built from source. That will allow for
building arbitrary non-released versions.

Change-Id: Ie177b8eb3c36c04bd4e2d5257c68e87b000cae85