Fix LLVM linker bug
The LLVM linker has a bug when linking together modules where the
destination module has two or more structurally identical struct types
and the source module also has these same struct types. The bug is that
all of the struct types in the source get mapped to an arbitrary struct
type in the destination. This can cause verification failures when
pointer types in call instructions don't match the argument types. E.g. in
Impala, TinyIntVal and BooleanVal get mixed up.

The LLVM linker has logic to match types by name when linking, but it did
not work correctly because of a bug in the hasType() method that incorrectly
reports that struct types were not present in the destination module: it
looks up a hash table that only contains the first structurally
identical type.

The fix is to maintain a hash table with all struct types in the
destination module and check that instead of the table with structurally
distinct types. The fix passes the LLVM test suite.