Build Kudu from source
With this change any arbitrary Kudu hash can be built. Previously only a
released version of Kudu could be used because the build was based on
parcels. The main use case is to upgrade the Kudu client/server to a
version that has yet to be released so compatibility can be tested
early.

The previous way of "building" Kudu was to repackage a Cloudera parcel.
For platforms that didn't have a parcel available, a stubbed client was
generated based on the symbols of a real client for a different
platform. The only consumer of the stub was the Impala project. Now
Impala will build the stub itself.

Change-Id: I2ec8494fb4e765ec796b31212c811af34e8514bd