Fix bug with publishing
A previous commit "Add support for Debug and Release+Asserts builds for
LLVM 3.7+" caused the native toolchain to publish artifacts with the
wrong versions and directories: the package name + version appeared
where previously there was only the package version. The commit
passed testing because the test was run with publishing disabled.

This fixes the problem. I did a test run with publishing enabled,
and the artifacts ended up in the right place on S3.

Change-Id: I8f14b43ada82f995e2385a27cfacfc6f23fd4a7c