Disable terminfo dependency for LLVM
LLVM uses libtinfo/libncurses to detect whether a terminal supports
colors. This is used to support Clang's colourful error messages.
However, if this feature is turned on, the way the LLVM libraries
are factored we have to bring in this dependency regardless of whether
we're actually making use of output colouring.

We don't want to depend on these libraries if building an embedded JIT,
so let's just disable the feature. The downside of this change is that
we won't get colourful error messages from toolchain clang.

Also remove the LLVM trunk build script that has been slowly bitrotting.