Add support for Kudu 1.0 RC and build additional OSes
Relaxes the requirements on the OSes that can be used to
build Kudu. This now allows building on Ubuntu 12,
Debian 7, and Debian 8. RHEL5, Debian 6, and SLES still fail
to build so the stub is produced.

Also removes some unnecessary kudu utilities (cfile-dump and
log-dump) that were being copied out of the build directory
because they have been removed at version 1.0. They were
never being used by Impala nor are related to building, so
it shouldn't matter.

Change-Id: I68c32b474464ccd7531b5c5bf163017cbb6f6f9b