IMPALA-4477: Bump Kudu to latest commit on master (60aa54e)
Update to Kudu git hash 60aa54e21b715a4e7a1931570c9736a2c7f211c6.

Also reverts a previous change that removed toolchain
binutils from the build PATH:
commit 2338fdd44a0439a27c6e26591c9598f859125442:
"Don't link with binutils from the toolchain"

The concern about building with the toolchain binutils was
that consumers of the toolchain binaries may need to use the
toolchain binutils, but Impala is the primary consumer and
is built with the toolchain binutils, so this doesn't seem
concerning. Further, consumers of the toolchain binaries may
also need to use other toolchain libraries and tools, so this
dependency makes sense.

This was tested in private builds which completed
successfully.

Change-Id: I506ce214010f9621b395e7a5ea80395d04a59df4