Allow overriding of Kudu github URL
This is a convenience to allow developers to build a branch of
kudu from a different github account (e.g. their own fork).

Testing:
Tested it locally on a private github enterprise repo. The kudu
build is still in progress but it successfully downloaded the
tarball and started building it.

Also confirmed that it still works if the URL is unspecified.

Change-Id: I181c599816435ca4997d84a6cd90386c3841ce42