IMPALA-4226, IMPALA-4227: bump max threads, handle dwz compressed symbols
IMPALA-4226: Increase the thread count limit in the minidump stack
resolver. The old limit to 4096 is arbitrary and we are hitting it from
time to time.

To test this I ran minidump_stackwalk on a minidump with more than 4096
threads.

IMPALA-4227: Add basic support for dwz dwarf extension
The dwz tool [1] can be used to compress symbols that occur in multiple
object files by moving them into a shared object file. It introduces new
DWARF macros to reference to those symbols.

Breakpad currently does not support those macros, which can lead to
crashes.

This change makes breakpad ignore these symbols. Impala's binaries don't
share symbols so we are not losing functionality with this change.

To test this I ran dump_syms on a binary compiled on CentOS 7, which
uses DWZ for symbol compression.

[1] https://sourceware.org/git/?p=dwz.git;a=summary

Change-Id: I1bf83edd8cda037c31a842801ad1445f3fd4f71e