IMPALA-5208, IMPALA-5187: Fixes for Breakpad #681, #728
IMPALA-5208 depends on Breakpad #681:
Breakpad had an issue [1] with random GUID generation on Linux, that
lead to processes started within the same second to write minidumps to
the same filename. Only one would win and overwrite minidumps of the
other processes.

IMPALA-5187 depends on Breakpad #728:
When writing a minidump on Linux, Breakpad called clone() in
linux/handler/exception_handler.cc with the CLONE_FILES flag. If the
parent process died while the child waited for the continuation signal,
the write side of the pipe 'fdes' stayed open in the child. The child
would not receive a SIGPIPE and would wait forever.

This change bumps the Breakpad version in the toolchain to include fixes
for the upstream bugs. It will allow us to remove the workaround we
introduced to address IMPALA-3794. I tested this by reverting the
workaround, removing any startup delay from the cluster startup scripts,
and then verified that the test_breakpad.py passes.

This change also removes the patch to increase the maximum number of
threads for minidump_stackwalk, which has been merged upstream.

[1] https://bugs.chromium.org/p/google-breakpad/issues/detail?id=681

Change-Id: If080919f0d241eac90edd8d23bf6007e09157e23