More GCC 7 fixes
My previous patch didn't fix all the build issues - I was building in a
pre-existing workspace with some packages already built with GCC 4.9.2.
This fixes the remaining compile issues.

This also adds LLVM 3.9, with the same patch we applied to LLVM 3.8
because it is needed to support the new GCC C++11 ABI. LLVM 3.8
remains the default.

Change-Id: Ibeaac85be762487de27019c2dac018d64e80f587