IMPALA-5477: Fix minidump-2-core tool in breakpad
The recent PPC port broke the minidump-2-core tool in the toolchain.
This change undoes a seemingly erroneous #ifdef to fix the tool on
non-PPC systems.

To test this, I built a toolchain locally and made sure that cores
created from minidumps with the tool can be opened by GDB.

I didn't test the change on PPC.

Change-Id: Iafe8cb59316bbec0a2c829f278a23a48238ad6e2