Bump thrift to 0.9.3-p2; Add bison 3.0.4
This patch upgrades thrift to 0.9.3-p2. Two patches are applied:
1. THRIFT-3650.
2. Remove Twisted Python dependency.

Thrift 0.9.3 depends on bison 2.5 or higher, which is not available on
some distributions including centos 5. This patch adds bison 3.0.4 to
the toolchain.
The minimum openssl version is bumped to 1.0.1. Bundled openssl will be
used to build thrift if the requirement is not met.
Testing: It builds on all supported OSes.

Change-Id: I710e977e4a182e5d3aaf58f1233c68c04585bebd