IMPALA-6377: Bump breakpad version to include fix for #752
This change bumps the Breakpad version in the toolchain to pull in the
fix for https://bugs.chromium.org/p/google-breakpad/issues/detail?id=752 .

It also copies the DWZ patch required for IMPALA-4227 to be applied to
the new version.

This change also removes the PPC patches from Breakpad, which I was not
able to get to apply cleanly. They will be re-applied in a later change.

Change-Id: I9b6212ab77eff2df0cb51339c25183ae4f22b07b