Bump libunwind version to 1.3-rc1
This changes bumps libunwind to version 1.3-rc1, the same version as in
Kudu. It also adds two patches to improve its interaction with asan and
tsan. The patch we had for 1.1 is not needed anymore.

It also adds a third patch to make the build and tests pass on CentOS 5.
The patch disables a test for remote unwinding. I wasn't able to figure
out why the test failed, but we currently only plan to use local
unwinding.

I tested this on my local dev environment with a prototype that rebases
our KRPC dependencies. I also tested this with a full build on a private
Jenkins job.

Change-Id: Ieb277ba55fb28145367d90dcbd868bf6b3c1710a
Reviewed-on: http://gerrit.cloudera.org:8080/10576
Reviewed-by: Lars Volker <lv@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>