Bump libunwind version to 1.3-rc1
This changes bumps libunwind to version 1.3-rc1, the same version as in
Kudu. It also adds two patches to improve its interaction with asan and
tsan. The patch we had for 1.1 is not needed anymore.

I tested this on my local dev environment with a prototype that rebases
our KRPC dependencies.

Change-Id: Ic34b78fd3272c12eb529ff06c7f9f1569d5def98
Reviewed-on: http://gerrit.cloudera.org:8080/10496
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>