Patch llvm to fix run-clang-tidy.py output
Clang's run-clang-tidy.py has some behaviors
that make it hard for Impala to get good output.
Specifically:
1. It does not synchronize output to stdout.
Output for different files can be interleaved
with each other and with the script's own output.
2. It does not suppress its own output when the
-quiet argument is specified. This output is hard
to separate from the actual output of clang-tidy.
3. It can silently skips files when clang-tidy
fails with an error code. This does not happen
with the existing set of clang-tidy checks, but
it could cause loss of coverage if a new check
caused clang-tidy to encounter an error.

This adds a patch to llvm that fixes these issues
in run-clang-tidy.py by synchronizing output,
respecting -quiet, and allowing an error from
clang-tidy to abort the run.

Since run-clang-tidy.py is not in the main llvm
source tree (it is in the extra package), our
existing method could not patch it. This changes
llvm to extract all of the archives into the
appropriate places before applying a unified
patch that can touch any of the source files from
any of the archives.

Testing:
 - Ran run-clang-tidy.py on normal Impala
   checkout without issues.
 - Added clang tidy checks that Impala
   currently doesn't pass and verified
   that the output with -quiet is
   sensible.

Change-Id: I9125cb0a908fd7005ee68aafb41f3afe93522632
Reviewed-on: http://gerrit.cloudera.org:8080/10788
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>