Patch Thrift to 0.9.3-p6 to eliminate erroneous ssl warnings.
This will be used to fix "IMPALA-8333: Remove Impala Shell warnings
part 2".

Remove misleading deprecation warning from the TSSLSocket.validate
method. This method is used to determine if the peer's SSL certificate
should be validated. The deprecation warning message is a mistake,
probably as the result of a confusion of the role of the validate method
with the use of a deprecated parameter, also named validate, that is
used in the initializer. This code was removed in subsequent Thrift
releases.

TESTING:

I built the toolchain locally and using jenkins.
I built impala with the new toolchain and tested that IMPALA-8333 is
fixed.

Change-Id: I84c2de4e6e3b21a4e3a5591f469d47ab34101687
Reviewed-on: http://gerrit.cloudera.org:8080/13383
Reviewed-by: Thomas Marshall <tmarshall@cloudera.com>
Tested-by: Thomas Marshall <tmarshall@cloudera.com>