Update cmake to 3.14.3
This updates to the latest cmake, which includes a few improvements that
may be helpful. In particular, the Ninja generator avoids generating
some false dependencies[1]

Instead of carrying over the patch to change one of the #defines in the
CMake source, this uses a newly-provided variable when building CMake
to accomplish the same.

I tested building Impala with this version of CMake on el7 and it worked
fine.

[1] https://gitlab.kitware.com/cmake/cmake/merge_requests/430

Change-Id: I1ad46a7b820bbb727ce47219f03f1d4a36301ab4
Reviewed-on: http://gerrit.cloudera.org:8080/13286
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>