Patch Avro to 1.7.4-p5 to add DATE type support.
This patch is needed because the Avro C library doesn't support
logical types yet. This is a quick-and-dirty patch to add DATE
type without a proper logical type implementation.

Note, that DECIMAL type support was added similarly in 1.7.4-p2.

This will be used in "IMPALA-8198: DATE: Read from avro."

TESTING:
I built the toolchain locally and using jenkins.
I built impala (including IMPALA-8198) with the new toolchain and
ran the core test suite.

Change-Id: I70aae9f16805e3ee806c39e9121b6525d83845b3
Reviewed-on: http://gerrit.cloudera.org:8080/13898
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>