IMPALA-8978: Fix Centos 6 docker image build issues
Everything needs TLSv1.2 now, and the version of nss that ships
with Centos 6.6 is too old. We already knew this and were
installing the nss from Centos 6.8, but we were installing it
after all the yum-install commands. The yum-install commands
now fail with TLS issues, so this installs the new nss before
invoking any other yum-install commands.

Separately, the download for ccache-3.3.3 is failing on Centos 6.
This failure is not detected, so it produces a docker image that
uses the built-in ccache. Unfortunately, there are bugs in that
version of ccache such that cmake does not compile properly. This
makes multiple changes to fix this:
1. This changes the URL to a working URL.
2. This changes postinstall.sh to run the downloads serially
   so that it does not miss the error code.
3. This adds a check for the ccache version in
   assert-dependencies-present.py

Testing:
 - Built all the docker images
 - Ran a toolchain build with the new docker images

Change-Id: I938a949a16eaf3291e765d67fbb1fd8912b8424f
Reviewed-on: http://gerrit.cloudera.org:8080/14366
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>