Fix prerequisites for building newer Kudu
Newer versions of Kudu require Java 8 to build, so all images need
to use Java 8. This was already true for everything except Debian8.
This switches the Debian8 docker container to use Java 8. This
required getting openjdk-8-jdk from jessie-backports. (Good blog
post on this: https://www.lucas-nussbaum.net/blog/?p=947 )
This adds a validation step to verify that java has version
1.8.0.

Separately, Kudu's thirdparty compilation of curl requires being
able to find krb5-config on the PATH. On SLES12, krb5-config
is in /usr/lib/mit/bin, so this adds that directory to PATH if
it exists.

Change-Id: Id4e776ea715ad7d13a978bc69c898261c1582384
Reviewed-on: http://gerrit.cloudera.org:8080/14892
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>