IMPALA-9265: Support for toolchain Kudu to provide Java artifacts
The build script was modified to generate Kudu JARs and add them
to the Kudu tarball.

redhat6 and redhat7 docker images were modified to update Java 8
to a newer version that is suitable for building the Java
artifacts.

ubuntu1404 docker image was modified to include CA certificate
file for Java.

Testing:
Ran the C6 toolchain build job to verify that native-toolchain
builds on all supported platforms.

Change-Id: Iba03dfe9c302513b825cbed7146c582e7d97c3af
Reviewed-on: http://gerrit.cloudera.org:8080/15072
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>