IMPALA-9837: Switch to GCC 7.5
This changes GCC from version 4.9.2 to 7.5.0. The libstdc++
ABI changed between 4.9.2, so packages compiled from this
version onward are not interchangeable with packages compiled
previously.

This also bumps the version of binutils from 2.26.1 to 2.28.

Now that the native toolchain is past the major C++11 ABI
change, it should be easier to explore using newer versions
of GCC. This adds standalone compilations of GCC 8.3 and
GCC 9.2.

Change-Id: Ia0a454c93a72b562ee46912cebbdc82b8b8375d2
Reviewed-on: http://gerrit.cloudera.org:8080/15205
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>