Add build script to create and publish Docker containers
Adds automation to build and publish the containers used by the
Docker-based toolchain builds. These can be used by the toolchain builds
to save time by using pre-created containers instead of building the
containers on the fly. They can also be used to provide a stable,
controlled environment for the toolchain builds.

The script can also be invoked from a Jenkins build environment.

Change-Id: Iac9b9b17b14c1aa8d9512437eb6ae01f5e896c55
Reviewed-on: http://gerrit.cloudera.org:8080/15675
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>