IMPALA-9686: Add readline to Python
Add the readline developer packages to the toolchain build containers.
These are needed by the Python build, as it figures out which packages
to build during compilation time by going through the environment and
finding these packages dynamically. Any packages that miss their runtime
dependencies are dropped from the build; the only sign is some warning
messages in the build log.

SLES12 puts ncurses include file in a separate subdirectory under
/usr/include. This breaks compiling readline on SLES12 if the extra
directoyr is not added to the include path, so fix this in Python's
build.sh

Runtime asserts were also added to the Python build driver to check
that the following modules were built successfully:
- readline
- ssl
- bz2

Tests:
- built the container images
- built the toolchain in the new container configs
- tested Python from the new build for the presence of readline support

Change-Id: I5bbe6fa16305a4462278306947535fad468cb3d8
Reviewed-on: http://gerrit.cloudera.org:8080/15831
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>