Clone the Kudu repo using HTTPS by default
Since GitHub started enforcing TLSv1.2 for HTTPS connections, the Kudu
repo on Github was cloned via SSH. Unfortunately the default SSH access
used for this depended on obscure default credentials, which failed
to work in some of the cases.
WIth Impala 4.x dropping support for older OS platforms, now all the
supported OSs are capable of TLSv1.2, so it is safe to reset default
access to HTTPS.

Tested by rebuilding the toolchain in a local Docker container; verified
that cloning the Kudu repo is successful with no user interaction
required.

Change-Id: I2486b34e7ae62eecb49ab0a602308b9fe898c8d0
Reviewed-on: http://gerrit.cloudera.org:8080/15832
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>