IMPALA-9985/IMPALA-10011: Update supported platforms, fix Maven download
Based on recent operating system releases and retirements the list of
platforms supported by the binary toolchain build is updated as follows:

Add Centos 8.
CentOS 8 support  starts with version 8.2, which is the most recent minor
release for this platform. Earlier releases were impacted by several
glibc problems, which impacted the GCC binaries in the toolchain (both
4.9.2 and 7.5), causing compilation failures when building Impala.
A few of these problems are listed in [1], all of which are fixed in
CentOS 8.2 by including an updated version of glibc.

Remove Debian 8.
Debian 8 has reached end-of-support on June 30th, 2020; see the
announcement at [2]. The repo signature verification problems described
in IMPALA-10011 were a corollary to this retirement, as infrastructure
fell into disuse.
This patch removes Debian 8 from the list of Docker containers built
for toolchain builds, it also removes Debian 8 from the toolchain target
list.

Additionally change the Maven download URL to point to the new Apache
official archive site for a stable and maintained repo location.

[1] https://issues.apache.org/jira/browse/IMPALA-9985#comment-17163097
[2] https://www.debian.org/News/2020/20200709

Tests:
1. Built all the containers for toolchain builds
2. Built and published a new build of the toolchain for the new list of
   platforms, using the containers prepared in step 1.
3. Ran Impala core-mode tests on Centos 7 and Centos 8. Impala compiled
   successfully, so this fixes IMPALA-9985. Tests passed on Centos 7.
   Centos 8 showed a single cipherset-related failure in
   thrift-server-test on Centos 8, for which an upstream Jira will be
   filed.

All these steps (except the thrift-server-test failure in step 3) were
successful; they were executed on private infrastructure.

Change-Id: Ib622f1e636e8b17b72c67b3d48964acea20db9ff
Reviewed-on: http://gerrit.cloudera.org:8080/16823
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>