IMPALA-10058: Use commit hash as version for Kudu java artifacts
This changes the Kudu build so that it no longer produces
jars with SNAPSHOT versions. Before this change, the Kudu
jars would have a version like 1.13.0-SNAPSHOT, and now
it is just the commit hash.

This should eliminate version conflicts where the solution
is to clear the .m2 cache.

This also bumps the Kudu commit hash to the latest. This
is needed because existing Kudu toolchain directories
have the old version format.

Testing:
 - Ran an Impala build with a toolchain produced by this

Change-Id: I352fdcb02ce3c11432cb5d6e7fdc564fe9b3c038
Reviewed-on: http://gerrit.cloudera.org:8080/16685
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>