IMPALA-9999: Build the toolchain with GCC 10
This switches the toolchain to build with GCC 10.
This required several version changes and patches
due to new warnings or changes in default settings:
1. LLVM is patched to fix compilation in lli
   (type mismatch on char)
2. crcutil is upgraded to the latest to handle the
   two digit GCC version. The new version incorporates
   the two patches previously used.
3. libunwind is upgraded to 1.5.1 due to GCC 10
   switch to -fno-common by default. The new
   version incorporates a couple of the previous
   patches.
4. Breakpad is upgraded to a recent commit (from late
   March 2022), which uses a newer version of lss
   that can compile with GCC 10.
5. Flatbuffers is upgraded to 1.12.0, which fixes
   a constness casting issue
6. The TPC-DS library is patched to use -fcommon,
   as GCC 10 switched to -fno-common by default.
7. GDB fails to compile on ARM due to the switch
   to -fno-common, so this adds newer GDB versions
   without that issue.

This upgrade GCC to 10.4 and binutils to 2.35.1.
This changes GCC's build command to build GCC using
link time optimization. This should speed up GCC's
execution, but it has no other impact. Building GCC
with LTO requires GCC compilation to use an updated
binutils, so this modifies the build script to build
binutils first and use that for GCC.

GCC 10.4 requires a patch to libstdc++ to fix missing
noexcepts, otherwise Clang compilation of Impala fails.

Testing:
 - This has built on all supported platforms, and
   Impala can build and operate with this toolchain.

Change-Id: I2fbfcdcb497ad60772bf0a7837e4043ae75bcfaa
Reviewed-on: http://gerrit.cloudera.org:8080/18806
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>