IMPALA-11004 (part 1): Upgrade glog library in toolchain
Third party libraries - Boost, glog and protobuf in Impala native
tool-chains are be older than the corresponding libraries used by Kudu.
We need to upgrade these three libraries before rebase be/src/kudu.

This patch upgrade glog from 0.3.4-p3 to 0.3.5-p3.
One of patches for glog 0.3.4 is merged to 0.3.5 so we need to apply
other two patches to 0.3.5.
Add one more patch which was picked by Kudu and it fixed symbolization
issue.

Testing:
 - Built native toolchain on local machine.
 - Built native toolchain with Jenkins job.
 - Built Impala with upgraded toolchain.

Change-Id: Ie6c57ea92b1eac03f7fd83962fed07fd28367e96
Reviewed-on: http://gerrit.cloudera.org:8080/17945
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Wenzhe Zhou <wzhou@cloudera.com>