IMPALA-11195: Disable SSL session renegotiation
This patch disables TLS ciphers renegotiation for TLSv1.2 and prior
protocol versions. Renegotiation is not possible in a TLSv1.3
connection.

In case of OpenSSL version 1.1.0h and newer, we are
using SSL_OP_NO_RENEGOTIATION option to disable all renegotiations. In
case of OpenSSL version prior to 1.1.0a, the undocumented flag
SSL3_FLAGS_NO_RENEGOTIATE_CIPHERS is used.

The moot point is the version interval between 1.1.0a and 1.1.0g
(inclusive): the SSL3_FLAGS_NO_RENEGOTIATE_CIPHERS flag is no longer
available from the application side, but SSL_OP_NO_RENEGOTIATION is not
yet present. So, if a server binary has been compiled with OpenSSL in
the specified version range, it's still advertising the renegotiation
option, even if it's run against OpenSSL 1.1.0h or later versions.

Change-Id: I44ee3ff729798834fdda5862f4e50dae8bb287a2
Reviewed-on: http://gerrit.cloudera.org:8080/18347
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>