IMPALA-10199: Add a toolchain build for Ubuntu 20
This adds support for building the toolchain
on Ubuntu 20.

With the new version of glibc in Ubuntu 20, GCC
and LLVM fail to compile due to an assert about
the offsets of particular fields in structs.
This adds a patch for GCC and LLVM to fix the
compilation error. These patches match the
changes made to libsanitizer in:
https://github.com/gcc-mirror/gcc/commit/4abc46b51af5751d657764d0c44b8a4aeed06302
This is the same libsanitizer code in both GCC
and LLVM code, so the patches are equivalent.

Testing:
 - Built the Ubuntu 20 docker image and built
   the toolchain.
 - Built with Ubutu 16 to check older versions

Change-Id: Ic3bb716448173d718a675b45bddd34228fdfac6d
Reviewed-on: http://gerrit.cloudera.org:8080/18505
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>