IMPALA-11383: Add Thrift 0.13.0, 0.14.2, and 0.16.0
This adds newer Thrift versions to address open
CVEs. Since upstream Impala now supports using
different versions of Thrift for different languages,
this includes multiple versions to allow flexibility.

Thrift 0.14.2 and 0.16.0 are suitable for any language.
Thrift 0.13.0 is not suitable for C++, because it does
not support overriding certain Thread methods that
Impala uses.

The newer version includes some patches that were
already on Thrift 0.11. The one new patch is a fix
for a SIGSEGV in Thrift's Thread destructor when
the Thread is uninitialized.

This also stops building the C++ fb303 code and
instead packages the fb303 thrift file. The fb303
library is not used by Impala's build and the C++
build for it is no longer reliably maintained.

This moves Thrift 0.9.0 and 0.9.3 to BUILD_HISTORICAL
as they are no longer in use.

Testing:
 - Ran full build

Change-Id: Ida147367c007dff3458508a477f7f6684305fb32
Reviewed-on: http://gerrit.cloudera.org:8080/18657
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>