IMPALA-11515: Do not use --load-average if ninja is installed
Kudu will use ninja if it is installed. Since ninja
does not support --load-average, the Kudu build will
fail if it runs on a machine with ninja installed.
This detects whether ninja is installed and does
not use --load-average. This only applies to one
step of the Kudu build. The usual method of building
native-toolchain does not have ninja installed, so
this is an unusual case.

Testing:
 - Tested the logic by hand inside a docker image. Tried
   without ninja installed, then installed ninja and tried
   again.

Change-Id: I7335186c70aa98de3c2b8d15264a29ec485ce784
Reviewed-on: http://gerrit.cloudera.org:8080/18957
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>