IMPALA-11605: Switch back to a patched flatbuffers 1.9.0
In IMPALA-9999, we moved to flatbuffers 1.12 to fix
a compilation issue of flatbuffers with newer GCC.
This requires Impala to use flatbuffers 1.12 for
Java as well, and that created a conflict with Hive.

This switches back to flatbuffers 1.9 (the same as
Hive) with a patch for the compilation issue.

Testing:
 - Built flatbuffers with GCC 10

Change-Id: Icd3b537ce7ffac61a311045453d4e063175a3300
Reviewed-on: http://gerrit.cloudera.org:8080/19032
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>