IMPALA-11603, IMPALA-12063: Address CVE in zlib by upgrading to 1.2.13
Zlib fixed CVE-2022-37434, an issue in inflateGetHeader() in 1.2.13.
This bumps the version of zlib to 1.2.13 to pick up this fix.

This also adds a build of Cloudflare zlib. Cloudflare zlib is a
drop-in replacement for the standard zlib library that has been
optimized to take advantage of SIMD and other processor support
on x86_64 and ARM. This adds a build of the latest Cloudflare zlib
as a new component. This version of Cloudflare zlib also contains
the fix for CVE-2022-37434.

Testing:
 - Ran a native-toolchain build

Change-Id: I14137848ebbe82f42df6a97fd24f5cdba4f65d21
Reviewed-on: http://gerrit.cloudera.org:8080/19748
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>