IMPALA-11949: Add AVRO C++ 1.11.1 version
This patch adds a new AVRO version 1.11.1 and new build files for AVRO
C++. Until the integration is finished, for backward compatibility, both
AVRO C and AVRO C++ will be part of the toolchain.

The upgrade is necessary to support logical types, which is required
for Iceberg DECIMAL types. The AVRO C library has not been meaningfully
developed for 3 years now and it does not support logical types,
meanwhile the C++ library is still under development and supports the
necessary types.

Change-Id: I307247cdf19ea8b9a97248624111bdaa05dd8496
Reviewed-on: http://gerrit.cloudera.org:8080/19546
Reviewed-by: Tamas Mate <tmater@apache.org>
Tested-by: Tamas Mate <tmater@apache.org>