IMPALA-11956: Update Python 3 version to 3.7.16
Python 3.7.4 is fairly old, even for Python 3.7 releases.
This is an issue for creating a Python 3 virtualenv,
because the virtualenv uses the bundled pip3 from the
toolchain Python. The bundled pip3 in Python 3.7.4
is old, and it no longer work properly with new
python packages. Upgrading gets a newer pip3 that works
properly.

Testing:
 - Built a Python 3 virtualenv locally

Change-Id: I7e2b6833bbefefa1e2a18c9900c7b123fe105d3c
Reviewed-on: http://gerrit.cloudera.org:8080/19562
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>