IMPALA-12069: Remove BUILD_HISTORICAL option
The BUILD_HISTORICAL option hasn't worked in a long time, and it
clutters the codebase with build code and patches for old
versions. This change removes the BUILD_HISTORICAL option and
all patches for old versions of components.

It also retires some old build logic for old versions. For
example, older versions of Snappy used autotools, but Snappy
switched to CMake. This cleans up the build script, removing
the old autotools logic. This applies to other components as
well.

This removes a couple locations of out of date OS-X darwin logic,
where darwin is referencing an old version of a component that
is clearly useless to modern Impala.

Testing:
 - Ran a toolchain build

Change-Id: Ib2df0af087af14f57e80b5ed94879bd1434501d4
Reviewed-on: http://gerrit.cloudera.org:8080/19755
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>