IMPALA-12070: Use autotools/bison from the OS rather than building it
This stops building autoconf, automake, bison, and libtool.
These utilities are already installed in the docker images,
so this removes any code that used the toolchain version
of these. This is useful from a license perspective, because
several of these are GPL licensed. Removing them avoids any
complications due to licensing, and it brings the native-toolchain
code closer to compatibility with Apache Impala.

This also removes unused build code for cyrus-sasl, krb5,
and openssl. This code remains in the history and can be
revived at any time.

Testing:
 - Ran a full toolchain build

Change-Id: I2f0603671fa443b26b3a339874f4154e853e2a06
Reviewed-on: http://gerrit.cloudera.org:8080/19756
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>