IMPALA-12100: Set REGISTER_INSTALL_PREFIX=Off for gflags build
When building the native-toolchain outside of Docker,
the Kudu build fails with link errors related to
glog's logging.cc having an undefined reference to
FlagRegisterer. This is because Kudu's glog build is
picking up Impala's gflags, which has a different signature
for FlagRegisterer.

It picks up Impala's gflags because gflags registers itself
in the user's home directory under the .cmake directory. Glog
must pick that up somehow. This is controlled by the
REGISTER_INSTALL_PREFIX CMake option for gflags. This sets
REGISTER_INSTALL_PREFIX=Off to avoid this interaction.

This also cleans up gflags' build.sh to remore the old non-CMake
code.

Testing:
 - Building the toolchain outside of Docker passed on x86_64
   with this change

Change-Id: I682ce40622f4bbad0c412de29006a3a9b4ae4765
Reviewed-on: http://gerrit.cloudera.org:8080/19798
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>