IMPALA-12202: Replace lsb_release with /etc/os-release
This switches the code in init.sh to use /etc/os-release
rather than lsb_release. /etc/os-release is available on
all supported Linux distributions, and it provides the
same information. There is a database of /etc/os-release
files at https://github.com/chef/os_release that helps
find the right comparisons.

This does not change existing Docker images to stop
installing lsb_release, because sometimes older
native-toolchain branches build on newer Docker images.
However, it does mean that future OSes added (like Redhat 9
and Ubuntu 22) will not need to provide lsb_release.
To support that use case, this adds the --no-lsb-rlease flag
to assert-dependencies-present.py, which tells it that
lsb_release is not required. Existing images do not specify
that flag, so lsb_release is still verified.

This removes the RELEASE_NAME environment variable and
obsolete code that used it. It also adds an assertion that
the utility 'cmp' is present. This is currently provided
as a transitive dependency of lsb_release, but any new
distribution without lsb_release will need to get it directly.

Testing:
 - Built new docker images and ran builds with them

Change-Id: I3aecaf0fddf27e2a8fbff00a772d6a1cdc6fd7eb
Reviewed-on: http://gerrit.cloudera.org:8080/20062
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>