IMPALA-12203: Default to Python 3 on modern distributions
This modifies the postinstall.sh script to set python to
point to python 3 for Linux distributions that do not
offer python 2. This will apply to Redhat 9 and Ubuntu 22
when support is added, but it does not apply to any
older distributions. The docker images can be used for
building older native-toolchain versions, so none of
the older distributions can be modified to use python 3.

To support this, this cleans up some old python logic
in a few packages. LLVM and Kudu can use system python
without any issues, and they both support python 3 for
this use case.

This removes some obsolete code from assert-dependencies-present.py
that used the Python 2 'platform' package. These checks
do not apply to any supported distributions.

This also changes postinstall.sh to use the last versions
of pip (20.3.4) and awscli (1.19.112) that support Python 2.
These are recent enough to also support newer Python 3 like
Python 3.10.

Testing:
 - Built Docker images and ran the build
 - Also ran the build for several older native-toolchain commits
   with the new Docker images to make sure that they
   continue to work.

Change-Id: Ib10589927509eeaa44b628709dcdb5e45c4dd034
Reviewed-on: http://gerrit.cloudera.org:8080/20063
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>