Omnibus: Add new versions of several libraries
To economize on toolchain builds, this adds new versions for
several libraries while keeping the existing version. This
allows us to test and upgrade the libraries without a new
toolchain build for each upgrade.

Here are the included new versions:
1. RE2 2023-03-01 (for IMPALA-12363)
2. ZLib 1.3.1
3. Cloudflare Zlib 7aa510344e
4. Gperftools 2.15
5. LZ4 1.9.4 (for IMPALA-12368)
6. ZSTD 1.5.5 (for IMPALA-12369)
7. Libunwind 1.8.1
8. CCTZ 2.3 and 2.4
9. jwt-cpp 0.6.0 and 0.7.0
10. Arrow 13.0.0
11. Mold 2.4.1
12. Python 3.8.18

Since Arrow and Mold have not been used by Impala yet, this removes
the code to build the old version. All others are additional builds
without removing the old version. These are small libraries, so
this has minimal impact on build time.

Testing:
 - Ran x86_64 / ARM toolchain builds

Change-Id: I2a900bc9b0c939391a3c6f970b1b5d99fbb42fcf
Reviewed-on: http://gerrit.cloudera.org:8080/21144
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>